pyproject.toml
[tool.poetry]
name = "label-studio-tools"
version = "0.1.0"
description = ""
authors = ["Kei Shimizu <suns.manage.k@gmail.com>"]
readme = "README.md"
packages = [{include = "label_studio_tools"}]

[tool.poetry.dependencies]
python = "^3.10.7"
pandas = "^2.0.3"
azure-storage-blob = "^12.17.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
